+++ internal/observability/event/event.go
@@ -16,6 +16,7 @@ type RequestInfo struct {
 	Method   string `json:"method,omitempty" class:"public"`
 	Path     string `json:"path,omitempty" class:"public"`
 	PublicId string `json:"public_id,omitempty" class:"public"`
+	ClientIp string `json:"client_ip,omitempty" class:"public"`
 }
 
 // UserInfo defines the fields captured about a user for a Boundary request.