+++ internal/servers/worker/proxy/proxy.go
@@ -15,6 +15,11 @@ import (
 // a provided ClientConn and the RemoteEndpoint, as well as the parameters to update
 // the connection in the connection repository.
 type Config struct {
+	// UserClientIp is the user's client IP
+	UserClientIp net.IP
+	// ClientAddress is the remote address (IP and port) of the client.  If
+	// there are any load balancers or proxies between the user and the worker,
+	// then it will be the address of the last one before the worker.
 	ClientAddress  *net.TCPAddr
 	ClientConn     *websocket.Conn
 	RemoteEndpoint string