+++ CHANGELOG.md
@@ -23,6 +23,8 @@ Canonical reference for changes, improvements, and bugfixes for Boundary.
 * host: Plugin-based host catalogs will now schedule updates for all
   of its host sets when its attributes are updated.
   ([PR](https://github.com/hashicorp/boundary/pull/1736))
+* controllers/workers: Add client IP to inbound request information which is included in
+  Boundary events ([PR](https://github.com/hashicorp/boundary/pull/1678))
 
 ## 0.7.1 (2021/11/18)